docs: extract the showcase component
